refactor: dissolve progress.rs into the render module
Pure code motion plus dead-code removal: the spinner styling moves to
render/style.rs (shared by the interactive renderer and the widget),
RollingLines joins its only users in render/mod.rs, and ProgressManager
becomes render/spinner.rs — the direct-use widget for commands that
render their own UX (network start/update). The ProgressBar trait and
execute_with_custom_progress existed for the deleted
MultiStepProgressBar and had one caller shape left, so
execute_with_progress now takes indicatif's ProgressBar directly.
