+++ crates/icp-cli/src/commands/network/update.rs
@@ -3,7 +3,7 @@ use std::sync::{Arc, OnceLock};
 use clap::Parser;
 use icp::{context::Context, network::managed::cache::download_launcher_version};
 
-use crate::progress::{ProgressManager, ProgressManagerSettings};
+use crate::render::{ProgressManager, ProgressManagerSettings};
 
 /// Update icp-cli-network-launcher to the latest version.
 #[derive(Parser, Debug)]
